backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
torch==2.1.0
torchvision==0.16.0
diffusers==0.21.4
transformers==4.35.0
accelerate==0.24.1
safetensors==0.4.0
controlnet-aux==0.0.10  # Fixed: 0.4.0 doesn't exist, using latest 0.0.x
# LLM dependencies
requests==2.31.0
openai>=1.0.0  # Supports both old and new API

